Use container/list instead of bytes.Buffer for the queue.
bytes.Buffer is really actually a simple byte slice.
The main problem with it is the fact that when you .Read() it, data doesn't actually get deleted. You get a "slice" of the slice, but the slice itself is never touched. Thus, if the userbase ever becomes big, it eventually becomes a problem because of the huge amount of unused bytes in the RAM.